% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{continentisland_ARG}
\alias{continentisland_ARG}
\title{Simulate ancestral recombination graph under continent-island model}
\usage{
continentisland_ARG(n = 2, loci = seq(0, 1000, 100), N = 100,
  rho = 0.001, mu = 0.01, generations = -1)
}
\arguments{
\item{n}{number of haploid samples.}

\item{loci}{genomic positions of loci.}

\item{N}{haploid population size.}

\item{rho}{recombination rate.}

\item{mu}{probability of migration from continent (aka mutation rate).}

\item{generations}{number of generations to simulate. If -1 then run until
all loci have coalesced (or until hard limit of 1e9 generation is reached).}
}
\description{
Simulate ancestral recombination graph under haploid
  continent-island model, which is equivalent to assuming infinite-alleles
  mutation.
}
