% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{continentisland_SMC_conditional_drop}
\alias{continentisland_SMC_conditional_drop}
\title{Drop genotype from conditional SMC under continent-island model}
\usage{
continentisland_SMC_conditional_drop(i = 1, recom, recom_time,
  event_time, coalescence, coalesce_target, migrate_target)
}
\arguments{
\item{i}{which genotype to drop.}

\item{recom}{boolean matrix specifying whether there was a recombination
between this locus and the previous locus.}

\item{recom_time}{matrix specifying recombination times.}

\item{event_time}{matrix of event timings.}

\item{coalescence}{boolean matrix specifying whether event is coalescence
(TRUE) or migration (FALSE).}

\item{coalesce_target}{matrix specifying, in the case of coalescence, which
other lineage to coalesce to.}

\item{migrate_target}{matrix specifying, in the case of migration, which
member of the continent population is the original ancestor.}
}
\description{
Drop genotype from conditional SMC under continent-island model.
}
