% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dummy1}
\alias{dummy1}
\title{Dummy function}
\usage{
dummy1(n = 10)
}
\arguments{
\item{n}{number of random draws}
}
\description{
Simple dummy function used to demonstrate linking between R and C++. Returns vector of draws from standard normal distribution.
}
\examples{
dummy1()
}
