% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{extract_polyIBD_summary}
\alias{extract_polyIBD_summary}
\title{Extract items from the output of \code{polyIBD::runMCMC} under the  \code{summary} list}
\usage{
extract_polyIBD_summary(input = ret, item = NULL)
}
\description{
TODO
}
