% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfilters.R
\name{genautocorr}
\alias{genautocorr}
\title{genautocorr calculates genetic autocorrelation for later linkage disequilibrium filtering}
\usage{
genautocorr(vcffile = NULL, vcfR = NULL)
}
\arguments{
\item{vcffile}{A variant call file (vcf) path. This VCF will be converted to an object of class \code{vcfR}.}
}
\description{
From an object of class \code{vcfR}, calculate the genetic autocorrelation as the estimate of linkage disequilibrium by genetic distance.
}
