% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfilters.R
\name{genautocorr_filter}
\alias{genautocorr_filter}
\title{polyIBD filter for linkage disequilibrium}
\usage{
genautocorr_filter(vcffile = NULL, vcfR = NULL, genautocorrresult = NULL,
  threshDist = 1000)
}
\arguments{
\item{vcffile}{A variant call file (vcf) path. This VCF will be converted to an object of class \code{vcfR}.}
}
\description{
Filtering an object of class \code{vcfR} for linkage disequilibrium via genetic autocorrelation.
}
