% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissioncalculationlookuptable.R
\name{Getemmissionlookuptable}
\alias{Getemmissionlookuptable}
\title{polyIBD Emission Calculation Look-up Table}
\usage{
Getemmissionlookuptable(m1max = 5, m2max = 5, popaffile)
}
\description{
.....
}
