% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getTransProbs}
\alias{getTransProbs}
\title{Get transition probabilities}
\usage{
getTransProbs(f, rho, zmax)
}
\arguments{
\item{f}{TODO}
}
\description{
Takes values of f, rho, and zmax. Produces rate matrix and calculates eigen values and vectors.
}
