% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{ggplot_IBDraster}
\alias{ggplot_IBDraster}
\title{Plot polyIBD outputs collectively}
\usage{
ggplot_IBDraster(x, trueIBD = NULL, truem1 = NULL, truem2 = NULL,
  truef = NULL, truek = NULL)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function \code{polyIBD::runMCMC}}

\item{trueIBD}{option to overlay a line corresponding to the true IBD (for example if using simulated data)}
}
\description{
Plots the MCMC chains for M1, M2, f, and k (generations), as well as the posterior distribution for the IBD matrix.
}
