% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_m1}
\alias{ggplot_m1}
\title{Trace ggplot of m1}
\usage{
ggplot_m1(x, ...)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function \code{polyIBD::runMCMC}}
}
\description{
Produces a simple MCMC trace ggplot2::geom_plot of the parameter \code{m1}, which represents the COI of the first sample.
}
