% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{ggplot_m2}
\alias{ggplot_m2}
\title{Trace ggplot of m2}
\usage{
ggplot_m2(x, ...)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function \code{polyIBD::runMCMC}}
}
\description{
Produces a simple MCMC trace ggplot2::geom_plot of the parameter \code{m2}, which represents the COI of the second sample.
}
