% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_rho}
\alias{ggplot_rho}
\title{Trace plot of rho}
\usage{
ggplot_rho(x, ...)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function \code{polyIBD::runMCMC}}
}
\description{
Produces a simple MCMC trace plot of the parameter \code{rho}, which represents the inverse of the average length of a recombinant block, and is a function of both the recombination rate and the number of generations separating the two lineages.
}
