% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot_SNPs_IBD}
\alias{ggplot_SNPs_IBD}
\title{Plot marginal IBD matrix with true IBD option and VCF that would have been generated for the two samples for comparison}
\usage{
ggplot_SNPs_IBD(x, trueIBD, snps)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function \code{polyIBD::runMCMC}}

\item{trueIBD}{option to overlay a line corresponding to the true IBD (for example if using simulated data)}
}
\description{
Plots the full posterior IBD matrix between two samples. SNPs are arranged in order on the x-axis, and the IBD level is given on the y-axis. This represents the number of genotypes that are shared between samples, with the lowest level (IBD=0) representing zero shared genotypes and hence no IBD. The intensity of colour represents the probability of each IBD level at each locus.
}
