% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{ggplot_trace}
\alias{ggplot_trace}
\title{ggtrace}
\usage{
ggplot_trace(data = NULL, mapping = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, inherit.aes = TRUE, ...)
}
\description{
ggtrace
}
