% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDSIMplotter.R
\name{IBDSIMplotter}
\alias{IBDSIMplotter}
\title{polyIBD Simulation Plotter -- IBD block plots}
\usage{
IBDSIMplotter(simDatareturn, runMCMCreturn)
}
\description{
This is a quick wrapper to plot the IBD block plot outputs from
}
