% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyIBDsimulator.R
\name{IBDsimulator2smplvcf}
\alias{IBDsimulator2smplvcf}
\title{polyIBD Empirical Simulator}
\usage{
IBDsimulator2smplvcf(n = 100, alpha = 0.1, beta = 0.1, MOIsmpl1 = 3,
  MOIsmpl2 = 4, m1max = 5, m2max = 5, fsim, rhosim, dsim = 1,
  contigs = "contig1")
}
\description{
Going off of empirical solution of model
}
