% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyIBDsimulator.R
\name{IBDsimulatorparams}
\alias{IBDsimulatorparams}
\title{polyIBD Empirical Simulator}
\usage{
IBDsimulatorparams(n = 100, shape1 = 0.1, shape2 = 0.1, MOIsmpl1 = 1,
  MOIsmpl2 = 1, fsim = 0.3, rhosim = 0.1, dsim = 1,
  contigs = "contig1")
}
\description{
Going off of empirical solution of model
}
