% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyIBDsimulator.R
\name{IBDsimulatorsimF_FixedAF}
\alias{IBDsimulatorsimF_FixedAF}
\title{polyIBD Empirical Simulator for f and a Fixed AF}
\usage{
IBDsimulatorsimF_FixedAF(n, p = 0.5, MOIsmpl1 = 1, MOIsmpl2 = 1,
  fsim = 0.3, rhosim = 0.1, dsim = 1, contigs = "contig1")
}
\description{
Going off of empirical solution of model
}
