% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTemissionprobSwitchFunction.R
\name{LikelihoodRowDeterminer}
\alias{LikelihoodRowDeterminer}
\title{polyIBD Likelihood Row Determiner -- switch function to determine which row of the likelihood/emission probabilities should be used}
\usage{
LikelihoodRowDeterminer(snpmatrixcolumn)
}
\description{
.....
}
