% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{markovchainIBDsim}
\alias{markovchainIBDsim}
\title{polyIBD Empirical markovchainIBD simulator}
\usage{
markovchainIBDsim(n, f, rho, d = 1)
}
\description{
Going off of empirical solution of model. Function inspired by Deonier's Computational Genomic Textbook?
}
