% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MCMC_VarMean}
\alias{MCMC_VarMean}
\title{MCMC_VarMean function}
\usage{
MCMC_VarMean(input)
}
\arguments{
\item{input}{a vector of numbers}
}
\description{
MCMC_VarMean function. return a markov chain
}
\examples{
MCMC_VarMean()
}
