% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissionprobabilityfunctions.R
\name{noshare_AaAa_emission}
\alias{noshare_AaAa_emission}
\title{polyIBD uncorrelated emission probability for A/A}
\usage{
noshare_AaAa_emission(p, q, m1, m2, z = 0)
}
\description{
.....
}
