% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissionprobabilityfunctions.R
\name{NoShareaAaEmission}
\alias{NoShareaAaEmission}
\title{polyIBD uncorrelated emission probability for a/Aa}
\usage{
NoShareaAaEmission(p, q, m1, m2, z = 0, e1 = 0.01, e2 = 0.01)
}
\description{
.....
}
