% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfcross_subset}
\alias{pfcross_subset}
\title{P. falciparum Cross Project Data}
\format{A vcf containing 4 samples and 100 biallelic SNPs}
\usage{
pfcross_subset
}
\description{
A variant call file (vcf) that contains SNPs on chromosome 1 for
the 3D7 and HB3 parents and the F1 progeny, C01 and C14.
}
\section{Warning}{
 If you copy and paste the code below, it will write to
                  your Desktop/temp directory (if it exists).
}

\section{Dependencies}{
 To generate this subsetted VCF, we used the
                       \code{NFBtools} package which is freely available
                       from GitHub with  \code{devtools::install_github("nickbrazeau/NFBtools")}
}

\section{Citation}{
 This VCF has generously been made publicly available by
the MalariaGEN Consortium and the P. falciparum Genetic Cross project led by
Alistair Miles. The manuscript is available through NCBI (PMID 27531718).
}

\keyword{datasets}
