% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.polyIBD}
\alias{plot.polyIBD}
\title{Default multi-panel plot for polyIBD class}
\usage{
\method{plot}{polyIBD}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function \code{polyIBD::runMCMC}}
}
\description{
The default \code{plot()} function can be called on objects returned from \code{polyIBD::runMCMC()} (which are of class \code{polyIBD}), to produce a multi-panel plot giving important MCMC output.
}
