% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_k}
\alias{plot_k}
\title{Trace plot of k}
\usage{
plot_k(x, ...)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function \code{polyIBD::runMCMC}}
}
\description{
Produces a simple MCMC trace plot of the parameter \code{k}, which represents the number of generations separating the two lineages (holding recombination rate constant).
}
