% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseplotting.R
\name{plot_m1}
\alias{plot_m1}
\title{Trace plot of m1}
\usage{
plot_m1(x, ...)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function
\code{polyIBD::runMCMC}}
}
\description{
Produces a simple MCMC trace plot of the parameter \code{m1},
which represents the COI of the first sample.
}
