% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_m2}
\alias{plot_m2}
\title{Trace plot of m2}
\usage{
plot_m2(x, ...)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function \code{polyIBD::runMCMC}}
}
\description{
Produces a simple MCMC trace plot of the parameter \code{m2}, which represents the COI of the second sample.
}
