% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotgenautocorr}
\alias{plotgenautocorr}
\title{Plot Genetic Autocorrelation result}
\usage{
plotgenautocorr(genautocorrobj)
}
\arguments{
\item{genautocorrobj}{an object of class \code{genautocorr}, as produced by the function \code{polyIBD::genautocorr}

do not export yet -- to do}
}
\description{
Plots the autocorrelation between loci (calculated from population allele frequencies)
}
