% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{polyIBD_iterations2tidy}
\alias{polyIBD_iterations2tidy}
\title{Extract burn-in and sampling iterations from the MCMC as part of the output from \code{polyIBD::runMCMC}}
\usage{
polyIBD_iterations2tidy(input = ret)
}
\description{
TODO
}
