% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotting.R
\docType{data}
\name{polyibdggtheme}
\alias{polyibdggtheme}
\title{polyibdggtheme}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 9.}
\usage{
polyibdggtheme
}
\description{
polyibdggtheme
}
\keyword{datasets}
