% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runMCMC2}
\alias{runMCMC2}
\title{Run polyIBD MCMC using Rcpp functions}
\usage{
runMCMC2(dat, SNP_dist, rho = 1, m_max = c(5, 5), burnin = 100,
  reps = 1000)
}
\arguments{
\item{dat}{TODO}
}
\description{
Run polyIBD MCMC using Rcpp functions.
}
\examples{
runMCMC2()
}
