% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissionprobabilityfunctions.R
\name{share_aa_emission}
\alias{share_aa_emission}
\title{polyIBD shared emission probability for a/a}
\usage{
share_aa_emission(p, q, m1, m2, z = 0, e1 = 0.01, e2 = 0.01)
}
\description{
.....
}
