% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissionprobabilityfunctions.R
\name{ShareAaAaEmission}
\alias{ShareAaAaEmission}
\title{polyIBD shared emission probability for Aa/Aa}
\usage{
ShareAaAaEmission(p, q, m1, m2, z = 0, e1 = 0.01, e2 = 0.01)
}
\description{
.....
}
