% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{simData}
\alias{simData}
\title{Simulate data from polyIBD model

TODO fix/update description}
\usage{
simData(pos = list(contig1 = sort(sample(1e+05, 100)), contig2 =
  sort(sample(1e+05, 100))), m1 = 1, m2 = 1, f = 0.5, rho = 1e-07,
  k = 5, p = NULL, p_shape1 = 0.1, p_shape2 = 0.1,
  propMissing = 0)
}
\arguments{
\item{pos}{TODO}

\item{m1}{TODO}

\item{m2}{TODO}

\item{f}{the average relatedness between the two samples}

\item{rho}{the recombination rate. TODO - units of this parameter}

\item{k}{the number of generations separating the two lineages}

\item{p}{TODO}

\item{p_shape1}{TODO}

\item{p_shape2}{TODO}

\item{propMissing}{TODO}
}
\description{
Walks along a vector of genomic locations and swiches between two states that
represent IBD and non-IBD using a Markov model. The parameters that dictate the chance of
switching state at any point include the average level of relatedness (\code{f}), the physical
distance between loci in units of base pairs (from \code{pos}), and the recombination rate
(\code{rho}), which is assumed constant over all loci.
}
