% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDataGeneratorBooster.R
\name{simDatGenerator}
\alias{simDatGenerator}
\title{Generate Sim Data for polyIBD}
\usage{
simDatGenerator(n, MOI1 = 2, MOI2 = 2, m1max = 5, m2max = 5, IBD,
  Contigs)
}
\description{
This is a very simple case that is too simple for real life...and is not what our model is estimating in expectation
}
