% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SMC_MCMC}
\alias{SMC_MCMC}
\title{Inference from conditional SMC under continent-island model}
\usage{
SMC_MCMC(x, n = 3, loci = seq(0, 1000, 100), N = 100, rho = 0.001,
  m = 0.01)
}
\arguments{
\item{x}{data in the form of a sequence of homo/het calls.}

\item{n}{complexity of infection of sample.}

\item{loci}{vector of genomic positions.}

\item{N}{haploid population size.}

\item{rho}{recombination rate.}

\item{m}{probability of migration from the continent per generation.}
}
\description{
Inference from conditional SMC under continent-island model
}
