% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfilters.R
\name{vcffilter}
\alias{vcffilter}
\title{polyIBD vcffilter}
\usage{
vcffilter(vcffile = NULL, formatGQ = 30, infoMQ = 50, infoQD = 25,
  infoSOR = 2, infoAF = 0.05, infoDP = NULL, prop.loci.missing = 0.05,
  biallelicSNPs = TRUE)
}
\arguments{
\item{vcffile}{A variant call file (vcf) path. This VCF will be converted to an object of class \code{vcfR}.}
}
\description{
Filtering a variant call file (VCF) using the vcfR package.
}
