% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{wrightfisher_SMC}
\alias{wrightfisher_SMC}
\title{Simulate from sequentially Markov coalescent under Wright-Fisher model}
\usage{
wrightfisher_SMC(n = 2, loci = seq(0, 1000, 100), N = 100,
  rho = 0.001)
}
\arguments{
\item{n}{number of haploid samples.}

\item{loci}{genomic positions of loci.}

\item{N}{haploid population size.}

\item{rho}{recombination rate.}

\item{generations}{number of generations to simulate. If -1 then run until
all loci have coalesced (or until hard limit of 1e9 generation is reached).}
}
\description{
Simulate from sequentially Markov coalescent approximation to
  the full ancestral recombination graph under haploid Wright-Fisher model. Note that because time is implemented as discrete this approximation will break down for large sample sizes,
}
