% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{wrightfisher_SMC_conditional}
\alias{wrightfisher_SMC_conditional}
\title{Simulate from conditional SMC under Wright-Fisher model}
\usage{
wrightfisher_SMC_conditional(n = 2, loci = seq(0, 1000, 100),
  N = 100, rho = 0.001, generations = NULL)
}
\arguments{
\item{n}{number of haploid samples.}

\item{loci}{genomic positions of loci.}

\item{N}{haploid population size.}

\item{rho}{recombination rate.}
}
\description{
Simulate from conditional SMC under Wright-Fisher model.
}
