% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{wrightfisher_SMC_naive}
\alias{wrightfisher_SMC_naive}
\title{Simulate from SMC under Wright-Fisher model by looping through
  discrete generations}
\usage{
wrightfisher_SMC_naive(n = 2, loci = seq(0, 1000, 100), N = 100,
  rho = 0.001, generations = 10)
}
\arguments{
\item{n}{number of haploid samples.}

\item{loci}{genomic positions of loci.}

\item{N}{haploid population size.}

\item{rho}{recombination rate.}

\item{generations}{number of generations to simulate. If -1 then run until
all loci have coalesced (or until hard limit of 1e9 generation is reached).}
}
\description{
Simulate from sequentially Markov coalescent (SMC) approximation
  to the full ancestral recombination graph (ARG) under haploid Wright-Fisher
  model. Simulate using naive method of looping through discrete generations.
}
