% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{continentisland_ARG}
\alias{continentisland_ARG}
\title{Simulate ancestral recombination graph under continent-island model}
\usage{
continentisland_ARG(n = 2, loci = seq(0, 1000, 100), N = 100,
  rho = 0.001, m = 0.01, generations = 100)
}
\arguments{
\item{n}{number of haploid samples.}

\item{loci}{genomic positions of loci.}

\item{N}{haploid population size.}

\item{rho}{recombination rate.}

\item{m}{probability of migration from continent (aka mutation rate).}

\item{generations}{number of generations to simulate.}
}
\description{
Simulate ancestral recombination graph under continent-island
  model. Assumes N haploid individuals each generation, with a probability m
  of originating from the continent each generation.
}
