% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{continentisland_SMC_conditional}
\alias{continentisland_SMC_conditional}
\title{Simulate from conditional SMC under continent-island model}
\usage{
continentisland_SMC_conditional(n = 2, loci = seq(0, 1000, 100),
  N = 100, rho = 0.001, mu = 0.01, generations = -1)
}
\arguments{
\item{n}{number of haploid samples.}

\item{loci}{genomic positions of loci.}

\item{N}{haploid population size.}

\item{rho}{recombination rate.}

\item{mu}{probability of migration from continent (aka mutation rate).}

\item{generations}{(dummy argument)}
}
\description{
Simulate from conditional SMC under continent-island model.
}
