% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionprobabilityfunctions.R
\name{getTransProbs}
\alias{getTransProbs}
\title{Calculate Transition Probabilities}
\usage{
getTransProbs(f, rho, zmax = 1, d = 1)
}
\arguments{
\item{f}{probability of IBD at equilibrium}

\item{rho}{rate of recombination}

\item{zmax}{maximum level of IBD possible. For simple binary IBD/not-IBD model use zmax=1}

\item{d}{genetic distance}
}
\description{
Take in model parameters and return transition probability matrix, evaluated at a given distance.
}
