% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{IBDsimulatorparams}
\alias{IBDsimulatorparams}
\title{polyIBD Empirical Simulator}
\usage{
IBDsimulatorparams(n = 100, m1 = 1, m2 = 1, f = 0.5, rho = 1,
  p = NULL, p_shape1 = 0.1, p_shape2 = 0.1, pos = 1:n)
}
\description{
Going off of empirical solution of model
}
