% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{IBDsimulatorparams}
\alias{IBDsimulatorparams}
\title{polyIBD Empirical Simulator}
\usage{
IBDsimulatorparams(n = 100, shape1 = 0.1, shape2 = 0.1, m1 = 1,
  m2 = 1, f = 0.5, rho = 1, d = 1, p = NULL, contigs = "contig1")
}
\description{
Going off of empirical solution of model
}
