% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyIBDsimulator.R
\name{markovchainIBDsim}
\alias{markovchainIBDsim}
\title{polyIBD Empirical markovchainIBD simulator}
\usage{
markovchainIBDsim(transprobs, n = n, fsim)
}
\description{
Going off of empirical solution of model
}
