% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseplotting.R
\name{plot_f}
\alias{plot_f}
\title{Trace plot of f posterior probability}
\usage{
plot_f(x, ...)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function \code{polyIBD::runMCMC}}
}
\description{
Produces a simple MCMC trace plot of the parameter \code{f}, which represents the posterior probability of identity by descent between two samples.
}
