% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rho}
\alias{plot_rho}
\title{Trace plot of rho}
\usage{
plot_rho(z)
}
\arguments{
\item{x}{an object of class \code{polyIBD}, as produced by the function \code{polyIBD::runMCMC}}
}
\description{
Produces a simple MCMC trace plot of the parameter \code{rho}, which represents the recombination rate.
}
