% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runMCMC}
\alias{runMCMC}
\title{Run polyIBD MCMC using Rcpp functions}
\usage{
runMCMC(input = NULL, m_max = 5, k_max = 10, rho = 1e-05,
  burnin = 100, samples = 1000, e1 = 0.05, e2 = 0.05,
  reportIteration = 1000)
}
\arguments{
\item{dat}{TODO}
}
\description{
Run polyIBD MCMC using Rcpp functions
}
