% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{simData}
\alias{simData}
\title{Simulate data from polyIBD model}
\usage{
simData(pos = list(contig1 = sort(sample(1e+05, 100)), contig2 =
  sort(sample(1e+05, 100))), m1 = 1, m2 = 1, f = 0.5, rho = 1e-07,
  k = 5, p = NULL, p_shape1 = 0.1, p_shape2 = 0.1,
  propMissing = 0)
}
\arguments{
\item{pos}{TODO}

\item{m1}{TODO}

\item{m2}{TODO}

\item{f}{TODO}

\item{rho}{TODO}

\item{k}{TODO}

\item{p}{TODO}

\item{p_shape1}{TODO}

\item{p_shape2}{TODO}

\item{propMissing}{TODO}
}
\description{
TODO
}
