% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2polyIBDinput.R
\name{vcf2polyIBDinput}
\alias{vcf2polyIBDinput}
\title{polyIBD vcf2polyIBDinput}
\usage{
vcf2polyIBDinput(vcffile = NULL, vcfR = NULL)
}
\arguments{
\item{vcffile}{is the file path that for the VCF. This VCF will be converted to an object of class \code{vcfR}.}

\item{vcfR}{is the file path that for the VCF. This VCF will be converted to an object of class \code{vcfR}.}
}
\description{
This function wraps the \code{vcfR} package to convert a vcffile or \code{vcfR} object to a SNP matrix suitable for polyIBD.
}
