% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{wrightfisher_ARG}
\alias{wrightfisher_ARG}
\title{Simulate ancestral recombination graph from Wright-Fisher model}
\usage{
wrightfisher_ARG(n, loci, N, rho, generations = 0)
}
\arguments{
\item{n}{number of haploid samples.}

\item{loci}{genomic positions of loci.}

\item{N}{(haploid) population size.}

\item{rho}{recombination rate.}

\item{max_time}{number of generations to simulate. If 0 then run until all
loci have coalesced.}
}
\description{
Simulate ancestral recombination graph from (haploid)
  Wright-Fisher model.
}
